refactor(suite-native): isolate portfolio graph visibility checks
Move portfolio graph header and time switch visibility decisions into the
components that render those UI pieces. PortfolioGraph stays responsible
for wiring the refetch handle and composing the graph surface, while each
child owns the selectors needed for its own visibility.

This preserves behavior and reduces rerenders caused by passing visibility
booleans through the portfolio graph container.
