+++ suite-native/module-home/src/screens/HomeScreen/components/PortfolioGraph.tsx
@@ -1,17 +1,7 @@
 import { forwardRef, useImperativeHandle } from 'react';
-import { useSelector } from 'react-redux';
 
-import { useAtomValue } from 'jotai';
-
-import { selectHasRunningDiscovery } from '@suite-common/wallet-core';
 import { VStack } from '@suite-native/atoms';
-import {
-    type RefetchPortfolioGraphParams,
-    portfolioGraphAtoms,
-    selectHasDeviceHistoryEnabledAccounts,
-    selectHasPortfolioGraphAccounts,
-    usePortfolioGraphData,
-} from '@suite-native/graph';
+import { type RefetchPortfolioGraphParams, usePortfolioGraphData } from '@suite-native/graph';
 
 import { IgnoredNetworksBanner } from './IgnoredNetworksBanner';
 import { PortfolioGraphTimeSwitch } from './PortfolioGraphTimeSwitch';
@@ -24,10 +14,6 @@ export type PortfolioGraphRef = {
 
 export const PortfolioGraph = forwardRef<PortfolioGraphRef>((_props, ref) => {
     const { refetchPortfolioGraph } = usePortfolioGraphData();
-    const hasDeviceHistoryEnabledAccounts = useSelector(selectHasDeviceHistoryEnabledAccounts);
-    const hasDeviceDiscovery = useSelector(selectHasRunningDiscovery);
-    const hasPortfolioGraphAccounts = useSelector(selectHasPortfolioGraphAccounts);
-    const isLoading = useAtomValue(portfolioGraphAtoms.isLoadingAtom);
 
     useImperativeHandle(
         ref,
@@ -37,16 +23,12 @@ export const PortfolioGraph = forwardRef<PortfolioGraphRef>((_props, ref) => {
         [refetchPortfolioGraph],
     );
 
-    const showHeader = hasPortfolioGraphAccounts || isLoading;
-
-    const showGraph = hasDeviceHistoryEnabledAccounts || hasDeviceDiscovery;
-
     return (
         <VStack spacing="sp24" testID="@home/portfolio/graph">
-            {showHeader && <PortfolioHeader />}
+            <PortfolioHeader />
             <PortfolioLineGraph refetchPortfolioGraph={refetchPortfolioGraph} />
             <IgnoredNetworksBanner />
-            {showGraph && <PortfolioGraphTimeSwitch />}
+            <PortfolioGraphTimeSwitch />
         </VStack>
     );
 });
