+++ suite-native/module-home/src/screens/HomeScreen/components/PortfolioGraphTimeSwitch.tsx
@@ -2,16 +2,18 @@ import { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { useServices } from '@suite-common/dependency-injection';
+import { selectHasRunningDiscovery } from '@suite-common/wallet-core';
 import { events, selectNativeAnalyticsDep } from '@suite-native/analytics';
 import {
     TimeSwitch,
     type TimeframeHoursValue,
+    selectHasDeviceHistoryEnabledAccounts,
     selectPortfolioGraphTimeframe,
     setPortfolioGraphTimeframe,
     timeSwitchItems,
 } from '@suite-native/graph';
 
-export const PortfolioGraphTimeSwitch = () => {
+const PortfolioGraphTimeSwitchContent = () => {
     const dispatch = useDispatch();
     const { analytics } = useServices(selectNativeAnalyticsDep);
     const timeframe = useSelector(selectPortfolioGraphTimeframe);
@@ -41,4 +43,13 @@ export const PortfolioGraphTimeSwitch = () => {
             onSelectTimeFrame={handleSelectPortfolioTimeframe}
         />
     );
+};
+
+export const PortfolioGraphTimeSwitch = () => {
+    const hasDeviceHistoryEnabledAccounts = useSelector(selectHasDeviceHistoryEnabledAccounts);
+    const hasDeviceDiscovery = useSelector(selectHasRunningDiscovery);
+
+    if (!hasDeviceHistoryEnabledAccounts && !hasDeviceDiscovery) return null;
+
+    return <PortfolioGraphTimeSwitchContent />;
 };
